+++ exporter/src/utils/PsdTextAttribute.cpp
@@ -17,6 +17,8 @@
 /////////////////////////////////////////////////////////////////////////////////////////////////
 
 #include "PsdTextAttribute.h"
+
+#include <algorithm>
 #include <array>
 #include <charconv>
 #include <iostream>
@@ -34,7 +36,7 @@ inline int HexCharToInt(const char c) {
   return result;
 }
 
-int PsdTextAttribute::stringFormatTransform(std::vector<char>& dst, const uint8_t* src,
+size_t PsdTextAttribute::stringFormatTransform(std::vector<char>& dst, const uint8_t* src,
                                             const int len) {
   dst.push_back('/');
   dst.push_back('0');
