+++ exporter/src/utils/PsdTextAttribute.h
@@ -17,6 +17,7 @@
 /////////////////////////////////////////////////////////////////////////////////////////////////
 
 #pragma once
+#include <memory>
 #include <string>
 #include <vector>
 
@@ -51,11 +52,11 @@ class PsdTextAttribute {
   void buildAttributeArray(int startPos, int endPos);
 
  private:
-  int stringFormatTransform(std::vector<char>& dst, const uint8_t* src, int len);
+  size_t stringFormatTransform(std::vector<char>& dst, const uint8_t* src, int len);
 
   std::vector<char> mem = {};
   char* src = nullptr;
-  int len = 0;
+  size_t len = 0;
   int pos = 0;
   int key = -1;
   PsdValueType valueType = PsdValueType::None;
